requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
openenv-core>=0.1.0
websockets>=12.0
